Added "heat up in a safe location" to the print_start marco
Added "heat up in a safe location" to the print_start marco. So the printer doesn't drip filament in the middle of the bed before the print start. So we'll move to the front left extreme and heat up there slightly above the bed. This is similar to how bambu heats up, where it moves to the metal plate to prevent oozing. Next mod will be adding this bambu ooze guard to the voron2.4